Add first draft of new tests
